svh_app/requirements.txt
# Core
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
python-multipart>=0.0.5,<0.1.0

# Database
sqlalchemy>=1.4.0,<2.0.0
databases[sqlite]>=0.5.0,<0.6.0
alembic>=1.7.0,<2.0.0

# Security
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.5,<0.1.0

# Data Validation
pydantic>=1.8.0,<2.0.0
email-validator>=1.1.3,<2.0.0

# Utilities
python-dotenv>=0.19.0,<0.20.0
python-dateutil>=2.8.2,<3.0.0
pytz>=2021.3,<2023.0

# Testing (optional, for development)
pytest>=6.2.5,<7.0.0
pytest-asyncio>=0.15.1,<0.16.0

# Documentation (optional, for development)
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0